Complete the first part of project
