+++ .babelrc
@@ -2,5 +2,5 @@
     "presets": [
         "@babel/preset-react"
     ],
-    "plugins": ["@babel/plugin-syntax-jsx"]
+    "plugins": ["@babel/plugin-proposal-private-property-in-object"]
   }
