+++ package.json
@@ -38,6 +38,7 @@
   "devDependencies": {
     "@babel/core": "^7.22.1",
     "@babel/eslint-parser": "^7.21.8",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@babel/plugin-syntax-jsx": "^7.21.4",
     "@babel/preset-react": "^7.22.3",
     "eslint": "^7.32.0",
