+++ src/App.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import './App.css';
+import Buttons from './components/Calculator';
+
+function App() {
+  return (
+    <div className="container">
+      <Buttons />
+    </div>
+  );
+}
+
+export default App;
